FIX: order of ssh commands and options in swarm node initialization
Fixed a bug where the reset of the docker daemon id by erasure of its
key.json file and restart would fail because of the order of the ssh
command arguments.

The ssh options, introduced by "-o", and the commands to execute were
mixed together. The reset of the docker daemon's id on the remote host
would then fail. Finally, the host would be added to the swarm,
potentially having a similar id to other nodes in the swarm and thus not
appearing as a different node when using the "docker info" command.