Remove test data files from repository
Removed the files used as upload payloads by the clients from the
repository to trim the fat.

In order to cope with the absence of some of these files, the client
scripts now check for the presence of the data file in the repository
before sending their first request.
If the data file is missing, the client creates it using the
gen_random_data script.