Fixed Shamir problem with files too big.
It still takes a lot of time to calculate the shamir secrets,
we will have to do really small tests for shamir or have no
timeout and wait until it is finished.

Also cleaned the docker file of py-proxy, going to add providers next.

Added logging both to the proxy and to the
pycoder.