Added dispatching of blocks among the different storage providers
The blocks generated by the encoders are now distributed between the
available storage providers in a simple/dumb round robin.

The element in charge of this distribution is the Dispatcher that keeps
track of the blocks location to be able to recover the files once they
are stored in the system.

py-proxy now uses this system to send the block for storage between
Dropbox, Google Drive and a local redis database.