Make storage providers used by dispatcher configurable
Changed the way the storage providers used by the dispatcher
configurable by adding a parameter to the Dispatcher initializer.

This change makes the use of multiple redis storage providers.

The current proxy implementation uses a JSON configuration file
(dispatcher.json) with a list of providers that can be used by the
dispatcher.