Make entanglement togglable in configuration
Add an entry in the dispatcher's configuration to indicate whether the
dispatcher should use block entanglement or not.

This change was made by changing the parameters received by the
Dispatcher's constructor from a list of providers (read from the
configuration file) to a dictionary with an entry for the entanglement
setting (entanglement) and a providers list under the key "providers".