FIX: Identifying a redis configuration in ProviderFactory
Fixed a bug where once the type of storage provider was detected in
ProviderFactory.get_provider, it was compared to the redis enum's value
using object comparison rather than value comparison (is vs ==).

This would result in a failure to identify the configuration as a redis
configuration eventually leading to one or multiple redis connections
trying directed at the wrong instance.