FIX: Dispatcher response on get requests
Fixed a bug where the dispatcher would always try to disentangle a list
of blocks whether it was set to do block entanglement or not.

This would result in a bug where the dispatcher would return a wrong
number of blocks for the pycoder to decode, making the pycoder and
eventually the HTTP request to the proxy fail.