Add possibility to replicate the entangled blocks and limit random block selection to data blocks
To complete the implementation of a Dagster-like entanglement scheme,
redundancy of data must be insured.
It is implemented here by the coder that returns copies of the
entangled data.

In order to avoid random dependencies between data and replica blocks, the
blocks returned by the GetRandomBlocks method in the `ProxyService` only
must be data blocks.
This precaution makes it easier to build reconstruction.