Change providers format in dispatcher.json
Changed the format of the providers section in dispatcher.json from a
list to a dictionary.
The key can then be used as a deterministic and human readable name for
the provider.
Fixed scripts/configure.py in accordance to change YAML serialization
library and freeze redis version in scripts/configure.py.