Change block fetching logic to consider replicas and better parallelism
Change the block fetching logic to iterate through the list of replicas
of a block in random order (for load balancing).

The block fetching is now fully parallelized by spawning one process
per block (when it used to spawn one thread per provider hosting one or
more block of the document we want to fetch) and going through all
available replicas (as explained above).