Optimize random block selection
The previous random selection implementation presented a problem as the
latency of the function would increase linearly with the size of the
block list.
To avoid this, we compute a list of random indexes within the list's
size and then only query the redis list for the elements we are interested in.

This leads to a significant performance improvement where the latency of
the method is no longer a function of the block index size but to the
number of elements to retrieve.