Make scripts/configure.py create volume directories for storage nodes
Make sure the volume directories are created because docker-compose
sometimes fails to do so.
This meant that we may end up with an incoherent state when we try to
restart playcloud after shutdown.
This fix should prevent this problem as long as scripts/configure.py is
run before starting playcloud.