Duplicate information about documents pointing to a block in the metadata
The `MetaBlock` data structure now has an extra field called
`entangled_with` that keeps track of the document that point to that
block for entanglement.
This makes the counting of the number of documents pointing to a block
 easier and prevents us from expensive lookups by scanning the entire
database.

The number of pointing documents is used as part of the
`has_been_entangled_enough` method in the files module as a way to make
sure that the block is protected enough and that its replicas can be
deleted.