Apply STeP offset to indices of missing fragments for reconstruction
When the proxy requests the reconstruction of a block, it is unaware of
the number source and pointer blocks used by the STeP configuration of
the coder.
In consequence it considers the `i^{th}` block it receives after encoding
to be in position `i` while it is actually in position `s + t + i` in
the original code.

In order to make sure that the proxy receives exactly the repaired
block it expects, the reconstruct method of the step entangler adds
an offset of `s + t` to the index of the missing fragment it receives
and then proceeds to reconstruction.