Use pointer reconstruction as last resort when reconstructing blocks
The script that scans and repairs blocks `pyproxy/repair.py` now first
scans for blocks to repair, groups them by document they belong to and
finally proceeds to reconstruction.

Looking at the number of missing blocks (or erasures), it tries to fix
them using RS.
If that is not possible, it falls back on reconstruction of the blocks
in codewords where they were used as pointers.